if new_path.exists(): return

+++ htignore/add-watermarks.py
@@ -1,6 +1,5 @@
 import asyncio, aiohttp, aiofiles
 from urllib.parse import quote
-import datetime
 import pathlib
 
 sem = asyncio.Semaphore(30)
@@ -16,18 +15,17 @@ async def fetch(path: str, session: aiohttp.ClientSession):
     p = pathlib.Path(path)
 
     if path.count('watermarked'):
-        # if (existing := pathlib.Path(path)).exists():
-        #    existing.unlink()
         return
     path = p
     if path.suffix not in ['.jpg', '.jpeg', '.png', '.webp', '.avif']: return
+    new_path = p.with_name(f"{path.stem}.watermarked{path.suffix}")
+    if new_path.exists(): return
     async with sem:
-        file_name = ('https://localhost/gallery/img-convert.' +
+        urlpath = ('https://localhost/gallery/img-convert.' +
                      f'php?img-path={quote(str(p.resolve()))}'
                      + f'&format=res{path.suffix}')
-        async with session.get(file_name) as resp:
+        async with session.get(urlpath) as resp:
             resp.raise_for_status()
-            new_path = p.with_name(f"{path.stem}.watermarked{path.suffix}")
             async with aiofiles.open(new_path, 'wb') as file:
                 await file.write(await resp.read())
         pass
@@ -44,7 +42,6 @@ async def main():
     pass
 
 
-# asyncio.run(main())
 try:
     asyncio.run(main())
 except* Exception as eg:
